+++ src/jsonyx/__init__.py
@@ -1,7 +1,5 @@
 # Copyright (C) 2024 Nice Zombies
 """Customizable JSON library for Python."""
-# TODO(Nice Zombies): add mapping_type
-# TODO(Nice Zombies): add seq_type
 # TODO(Nice Zombies): update badge branch
 from __future__ import annotations
 
@@ -129,12 +127,18 @@ def read(
     filename: _StrPath,
     *,
     allow: Container[str] = NOTHING,
+    mapping_type: type = dict,
+    seq_type: type = list,
     use_decimal: bool = False,
 ) -> Any:
     """Deserialize a JSON file to a Python object.
 
+    .. versionchanged:: 2.0 Added ``mapping_type`` and ``seq_type``.
+
     :param filename: the path to the JSON file
     :param allow: the allowed JSON deviations
+    :param mapping_type: the mapping type
+    :param seq_type: the sequence type
     :param use_decimal: use :class:`decimal.Decimal` instead of :class:`float`
     :raises JSONSyntaxError: if the JSON file is invalid
     :return: a Python object.
@@ -149,20 +153,31 @@ def read(
     ...
     ['filesystem API']
     """
-    return Decoder(allow=allow, use_decimal=use_decimal).read(filename)
+    return Decoder(
+        allow=allow,
+        mapping_type=mapping_type,
+        seq_type=seq_type,
+        use_decimal=use_decimal,
+    ).read(filename)
 
 
 def load(
     fp: _SupportsRead[bytes | str],
     *,
     allow: Container[str] = NOTHING,
+    mapping_type: type = dict,
+    seq_type: type = list,
     root: _StrPath = ".",
     use_decimal: bool = False,
 ) -> Any:
     """Deserialize an open JSON file to a Python object.
 
+    .. versionchanged:: 2.0 Added ``mapping_type`` and ``seq_type``.
+
     :param fp: an open JSON file
     :param allow: the allowed JSON deviations
+    :param mapping_type: the mapping type
+    :param seq_type: the sequence type
     :param root: the path to the archive containing this JSON file
     :param use_decimal: use :class:`decimal.Decimal` instead of :class:`float`
     :raises JSONSyntaxError: if the JSON file is invalid
@@ -174,7 +189,12 @@ def load(
     >>> json.load(io)
     ['streaming API']
     """
-    return Decoder(allow=allow, use_decimal=use_decimal).load(fp, root=root)
+    return Decoder(
+        allow=allow,
+        mapping_type=mapping_type,
+        seq_type=seq_type,
+        use_decimal=use_decimal,
+    ).load(fp, root=root)
 
 
 def loads(
@@ -182,13 +202,19 @@ def loads(
     *,
     allow: Container[str] = NOTHING,
     filename: _StrPath = "<string>",
+    mapping_type: type = dict,
+    seq_type: type = list,
     use_decimal: bool = False,
 ) -> Any:
     """Deserialize a JSON string to a Python object.
 
+    .. versionchanged:: 2.0 Added ``mapping_type`` and ``seq_type``.
+
     :param s: a JSON string
     :param allow: the allowed JSON deviations
     :param filename: the path to the JSON file
+    :param mapping_type: the mapping type
+    :param seq_type: the sequence type
     :param use_decimal: use :class:`decimal.Decimal` instead of :class:`float`
     :raises JSONSyntaxError: if the JSON string is invalid
     :return: a Python object
@@ -199,9 +225,12 @@ def loads(
 
     .. tip:: Specify ``filename`` to display the filename in error messages.
     """
-    return Decoder(allow=allow, use_decimal=use_decimal).loads(
-        s, filename=filename,
-    )
+    return Decoder(
+        allow=allow,
+        mapping_type=mapping_type,
+        seq_type=seq_type,
+        use_decimal=use_decimal,
+    ).loads(s, filename=filename)
 
 
 def write(